+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.15 - TBD
 
+Features
+
+- Added the `-CLNK` compiler option to specify a custom command to provide as a linker option. For example: `"-lcurl"`.
+
 Bug Fixes
 
 - Fixed the division (`/`) operator to ensure exact arguments will produce inexact results rather than incorrect fixnum results. The example given in R7RS is that `(/ 3 4)` must not return a mathematically incorrect result of `0`. This was previously broken in compiled code though `icyc` would return the correct result.